💄 (#13): Integraiting simple analysis into UI

+++ MaxTemp.Logic/ITemperatureService.cs
@@ -6,6 +6,6 @@ public interface ITemperatureService
 {
     public TemperatureEntry GetMaxForSensor(string sensorId);
     public TemperatureEntry GetMinForSensor(string sensorId);
-    public double GetAverageForSensor(string sensorId);
+    public decimal GetAverageForSensor(string sensorId);
 
 }
